Remove workaround for overlays on iOS devices
PR https://github.com/Collaborne/paper-fab-speed-dial/pull/2 introduces a workaround to not show the overlay on iOS devices.
This workaround seems to be no longer necessary with the latest Polymer version and Polymer element dependencies on Safari of a recent iOS version (9.3.5).
The workaround gets removed for consistent behavior across devices and to reduce code complexity.